Clean up browser logs helper ownership
