+++ tests/Aspire.Hosting.Tests/BrowserLogsSessionManagerTests.cs
@@ -3,7 +3,6 @@
 
 using System.Net;
 using System.Net.Sockets;
-using System.Net.WebSockets;
 using System.Text;
 using Aspire.Hosting.Tests.Utils;
 using Microsoft.Extensions.Logging;
@@ -477,68 +476,6 @@ public class BrowserLogsSessionManagerTests
     }
 
     [Fact]
-    public void TrySelectTrackedTargetId_PrefersUnattachedBlankPage()
-    {
-        var targetId = BrowserLogsRunningSession.TrySelectTrackedTargetId(
-        [
-            new BrowserLogsTargetInfo { TargetId = "restored-page", Type = "page", Url = "https://example.com", Attached = false },
-            new BrowserLogsTargetInfo { TargetId = "service-worker", Type = "service_worker", Url = "https://example.com/sw.js", Attached = false },
-            new BrowserLogsTargetInfo { TargetId = "launcher-page", Type = "page", Url = "about:blank", Attached = false }
-        ]);
-
-        Assert.Equal("launcher-page", targetId);
-    }
-
-    [Fact]
-    public void TrySelectTrackedTargetId_FallsBackToFirstUnattachedPage()
-    {
-        var targetId = BrowserLogsRunningSession.TrySelectTrackedTargetId(
-        [
-            new BrowserLogsTargetInfo { TargetId = "attached-page", Type = "page", Url = "about:blank", Attached = true },
-            new BrowserLogsTargetInfo { TargetId = "fallback-page", Type = "page", Url = "chrome://newtab/", Attached = false }
-        ]);
-
-        Assert.Equal("fallback-page", targetId);
-    }
-
-    [Fact]
-    public async Task BrowserConnectionDiagnosticsLogger_LogsConnectionProblems()
-    {
-        var resourceLoggerService = ConsoleLoggingTestHelpers.GetResourceLoggerService();
-        var resourceName = "web-browser-logs";
-        var diagnostics = new BrowserConnectionDiagnosticsLogger("session-0001", resourceLoggerService.GetLogger(resourceName));
-
-        var logs = await CaptureLogsAsync(resourceLoggerService, resourceName, targetLogCount: 4, () =>
-        {
-            diagnostics.LogSetupFailure(
-                "Setting up the tracked browser debug connection",
-                new InvalidOperationException("Connecting to the tracked browser debug endpoint failed.", new TimeoutException("Timed out waiting for a tracked browser protocol response to 'Target.attachToTarget'.")));
-            diagnostics.LogConnectionLost(
-                new InvalidOperationException("Browser debug connection closed by the remote endpoint with status 'EndpointUnavailable' (1001): browser crashed"));
-            diagnostics.LogReconnectAttemptFailed(
-                2,
-                new InvalidOperationException("Attaching to the tracked browser target failed.", new TimeoutException("Timed out waiting for a tracked browser protocol response to 'Target.attachToTarget'.")));
-            diagnostics.LogReconnectFailed(
-                new InvalidOperationException("Connecting to the tracked browser debug endpoint failed.", new WebSocketException("Connection refused")));
-        });
-
-        Assert.Collection(
-            logs,
-            log => Assert.Equal(
-                "2000-12-29T20:59:59.0000000Z [session-0001] Setting up the tracked browser debug connection failed: InvalidOperationException: Connecting to the tracked browser debug endpoint failed. --> TimeoutException: Timed out waiting for a tracked browser protocol response to 'Target.attachToTarget'.",
-                log.Content),
-            log => Assert.Equal(
-                "2000-12-29T20:59:59.0000000Z [session-0001] Tracked browser debug connection lost: InvalidOperationException: Browser debug connection closed by the remote endpoint with status 'EndpointUnavailable' (1001): browser crashed. Attempting to reconnect.",
-                log.Content),
-            log => Assert.Equal(
-                "2000-12-29T20:59:59.0000000Z [session-0001] Reconnect attempt 2 failed: InvalidOperationException: Attaching to the tracked browser target failed. --> TimeoutException: Timed out waiting for a tracked browser protocol response to 'Target.attachToTarget'.",
-                log.Content),
-            log => Assert.Equal(
-                 "2000-12-29T20:59:59.0000000Z [session-0001] Unable to reconnect tracked browser debug connection. Closing the tracked browser session. Last error: InvalidOperationException: Connecting to the tracked browser debug endpoint failed. --> WebSocketException: Connection refused",
-                 log.Content));
-    }
-
-    [Fact]
     public async Task StartSessionAsync_ThrowsWhenManagerIsDisposing()
     {
         var sessionFactory = new ThrowIfCalledSessionFactory();
@@ -565,9 +502,6 @@ public class BrowserLogsSessionManagerTests
 
         Assert.False(sessionFactory.WasCalled);
     }
-
-    private static Task<IReadOnlyList<LogLine>> CaptureLogsAsync(ResourceLoggerService resourceLoggerService, string resourceName, int targetLogCount, Action writeLogs) =>
-        ConsoleLoggingTestHelpers.CaptureLogsAsync(resourceLoggerService, resourceName, targetLogCount, writeLogs);
 
     private static void WithTempUserDataDirectory(Action<string> action)
     {
